GEODE-7994: Refactor Native Redis Acceptance Tests
