+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/GeoNativeRedisAcceptanceTest.java
@@ -14,10 +14,10 @@
  */
 package org.apache.geode.redis;
 
-
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
+import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.rules.TestRule;
 import org.testcontainers.containers.GenericContainer;
@@ -27,7 +27,7 @@ import org.apache.geode.test.junit.categories.RedisTest;
 import org.apache.geode.test.junit.rules.IgnoreOnWindowsRule;
 
 @Category({RedisTest.class})
-public class ListsDockerAcceptanceTest extends ListsIntegrationTest {
+public class GeoNativeRedisAcceptanceTest extends GeoIntegrationTest {
 
   // Docker compose does not work on windows in CI. Ignore this test on windows
   // Using a RuleChain to make sure we ignore the test before the rule comes into play
@@ -44,5 +44,42 @@ public class ListsDockerAcceptanceTest extends ListsIntegrationTest {
   @AfterClass
   public static void tearDown() {
     jedis.close();
+  }
+
+  // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+
+  @Test
+  public void testGeoHash() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testGeoRadiusByMemberWithCoord() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testGeoRadiusByMemberFull() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testGeoRadiusByMemberBasic() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testGeoRadiusByMemberNorth() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testGeoPos() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testGeoRadiusByMemberWithDist() {
+    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
   }
 }
