+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/HashesNativeRedisAcceptanceTest.java
@@ -14,10 +14,11 @@
  */
 package org.apache.geode.redis;
 
+import java.util.Random;
+
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
-import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.rules.TestRule;
 import org.testcontainers.containers.GenericContainer;
@@ -27,7 +28,7 @@ import org.apache.geode.test.junit.categories.RedisTest;
 import org.apache.geode.test.junit.rules.IgnoreOnWindowsRule;
 
 @Category({RedisTest.class})
-public class GeoDockerAcceptanceTest extends GeoIntegrationTest {
+public class HashesNativeRedisAcceptanceTest extends HashesIntegrationTest {
 
   // Docker compose does not work on windows in CI. Ignore this test on windows
   // Using a RuleChain to make sure we ignore the test before the rule comes into play
@@ -38,48 +39,14 @@ public class GeoDockerAcceptanceTest extends GeoIntegrationTest {
   public static void setUp() {
     GenericContainer redisContainer = new GenericContainer<>("redis:5.0.6").withExposedPorts(6379);
     redisContainer.start();
+    rand = new Random();
     jedis = new Jedis("localhost", redisContainer.getFirstMappedPort(), 10000000);
+    jedis2 = new Jedis("localhost", redisContainer.getFirstMappedPort(), 10000000);
   }
 
   @AfterClass
   public static void tearDown() {
     jedis.close();
-  }
-
-  // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-
-  @Test
-  public void testGeoHash() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-  }
-
-  @Test
-  public void testGeoRadiusByMemberWithCoord() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-  }
-
-  @Test
-  public void testGeoRadiusByMemberFull() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-  }
-
-  @Test
-  public void testGeoRadiusByMemberBasic() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-  }
-
-  @Test
-  public void testGeoRadiusByMemberNorth() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-  }
-
-  @Test
-  public void testGeoPos() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
-  }
-
-  @Test
-  public void testGeoRadiusByMemberWithDist() {
-    // TODO: See JIRA GEODE-7909 Update Geo* commands in Geode Redis to match native Redis
+    jedis2.close();
   }
 }
