+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/PersistNativeRedisAcceptanceTest.java
@@ -23,12 +23,12 @@ import org.junit.rules.TestRule;
 import org.testcontainers.containers.GenericContainer;
 import redis.clients.jedis.Jedis;
 
-import org.apache.geode.redis.general.ExpireAtIntegrationTest;
+import org.apache.geode.redis.general.PersistIntegrationTest;
 import org.apache.geode.test.junit.categories.RedisTest;
 import org.apache.geode.test.junit.rules.IgnoreOnWindowsRule;
 
 @Category({RedisTest.class})
-public class ExpireAtDockerAcceptanceTest extends ExpireAtIntegrationTest {
+public class PersistNativeRedisAcceptanceTest extends PersistIntegrationTest {
 
   // Docker compose does not work on windows in CI. Ignore this test on windows
   // Using a RuleChain to make sure we ignore the test before the rule comes into play
@@ -40,11 +40,13 @@ public class ExpireAtDockerAcceptanceTest extends ExpireAtIntegrationTest {
     GenericContainer redisContainer = new GenericContainer<>("redis:5.0.6").withExposedPorts(6379);
     redisContainer.start();
     jedis = new Jedis("localhost", redisContainer.getFirstMappedPort(), REDIS_CLIENT_TIMEOUT);
+    jedis2 = new Jedis("localhost", redisContainer.getFirstMappedPort(), REDIS_CLIENT_TIMEOUT);
   }
 
   @AfterClass
-  public static void classLevelTearDown() {
+  public static void tearDown() {
     jedis.close();
+    jedis2.close();
   }
 
 }
