+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/PubSubNativeRedisAcceptanceTest.java
@@ -12,7 +12,9 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
+
 package org.apache.geode.redis;
+
 
 
 import org.junit.AfterClass;
@@ -27,7 +29,9 @@ import org.apache.geode.test.junit.categories.RedisTest;
 import org.apache.geode.test.junit.rules.IgnoreOnWindowsRule;
 
 @Category({RedisTest.class})
-public class SetsDockerAcceptanceTest extends SetsIntegrationTest {
+public class PubSubNativeRedisAcceptanceTest extends PubSubIntegrationTest {
+
+  private static GenericContainer redisContainer;
 
   // Docker compose does not work on windows in CI. Ignore this test on windows
   // Using a RuleChain to make sure we ignore the test before the rule comes into play
@@ -36,15 +40,19 @@ public class SetsDockerAcceptanceTest extends SetsIntegrationTest {
 
   @BeforeClass
   public static void setUp() {
-    GenericContainer redisContainer = new GenericContainer<>("redis:5.0.6").withExposedPorts(6379);
+    redisContainer = new GenericContainer<>("redis:5.0.6").withExposedPorts(6379);
     redisContainer.start();
-    jedis = new Jedis("localhost", redisContainer.getFirstMappedPort(), 10000000);
-    jedis2 = new Jedis("localhost", redisContainer.getFirstMappedPort(), 10000000);
+    subscriber = new Jedis("localhost", redisContainer.getFirstMappedPort(), REDIS_CLIENT_TIMEOUT);
+    publisher = new Jedis("localhost", redisContainer.getFirstMappedPort(), REDIS_CLIENT_TIMEOUT);
   }
 
   @AfterClass
   public static void tearDown() {
-    jedis.close();
-    jedis2.close();
+    subscriber.close();
+    publisher.close();
+  }
+
+  public int getPort() {
+    return redisContainer.getFirstMappedPort();
   }
 }
