+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/RenameNativeRedisAcceptanceTest.java
@@ -15,20 +15,24 @@
 
 package org.apache.geode.redis;
 
+
+import java.util.Random;
+
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
+import org.junit.Ignore;
+import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.rules.TestRule;
 import org.testcontainers.containers.GenericContainer;
 import redis.clients.jedis.Jedis;
 
-import org.apache.geode.redis.general.ExpireIntegrationTest;
 import org.apache.geode.test.junit.categories.RedisTest;
 import org.apache.geode.test.junit.rules.IgnoreOnWindowsRule;
 
 @Category({RedisTest.class})
-public class ExpireDockerAcceptanceTest extends ExpireIntegrationTest {
+public class RenameNativeRedisAcceptanceTest extends RenameIntegrationTest {
 
   private static GenericContainer redisContainer;
 
@@ -39,14 +43,32 @@ public class ExpireDockerAcceptanceTest extends ExpireIntegrationTest {
 
   @BeforeClass
   public static void setUp() {
+    rand = new Random();
     redisContainer = new GenericContainer<>("redis:5.0.6").withExposedPorts(6379);
     redisContainer.start();
     jedis = new Jedis("localhost", redisContainer.getFirstMappedPort(), REDIS_CLIENT_TIMEOUT);
   }
 
   @AfterClass
-  public static void classLevelTearDown() {
+  public static void tearDown() {
     jedis.close();
   }
 
+  public int getPort() {
+    return redisContainer.getFirstMappedPort();
+  }
+
+  @Test
+  public void testSortedSet() {
+    // TODO: GEODE-7910 Update RENAME command in Geode Redis to match native Redis
+  }
+
+  @Test
+  public void testList() {
+    // TODO: GEODE-7910 Update RENAME command in Geode Redis to match native Redis
+  }
+
+  @Test
+  @Ignore("Test only applies to Geode Redis, ignored for native Redis")
+  public void testProtectedString() {}
 }
