+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/SetsNativeRedisAcceptanceTest.java
@@ -14,6 +14,7 @@
  */
 package org.apache.geode.redis;
 
+
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
@@ -26,7 +27,7 @@ import org.apache.geode.test.junit.categories.RedisTest;
 import org.apache.geode.test.junit.rules.IgnoreOnWindowsRule;
 
 @Category({RedisTest.class})
-public class SortedSetsDockerAcceptanceTest extends SortedSetsIntegrationTest {
+public class SetsNativeRedisAcceptanceTest extends SetsIntegrationTest {
 
   // Docker compose does not work on windows in CI. Ignore this test on windows
   // Using a RuleChain to make sure we ignore the test before the rule comes into play
@@ -38,10 +39,12 @@ public class SortedSetsDockerAcceptanceTest extends SortedSetsIntegrationTest {
     GenericContainer redisContainer = new GenericContainer<>("redis:5.0.6").withExposedPorts(6379);
     redisContainer.start();
     jedis = new Jedis("localhost", redisContainer.getFirstMappedPort(), 10000000);
+    jedis2 = new Jedis("localhost", redisContainer.getFirstMappedPort(), 10000000);
   }
 
   @AfterClass
   public static void tearDown() {
     jedis.close();
+    jedis2.close();
   }
 }
