+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/HashesIntegrationTest.java
@@ -796,7 +796,6 @@ public class HashesIntegrationTest {
   @After
   public void flushAll() {
     jedis.flushAll();
-    jedis2.flushAll();
   }
 
   @AfterClass
