+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/StringsIntegrationTest.java
@@ -82,7 +82,6 @@ public class StringsIntegrationTest {
   @After
   public void flushAll() {
     jedis.flushAll();
-    jedis2.flushAll();
   }
 
   @AfterClass
